1、/reading-rank返回学习行为分值 5./system-statistics/student-overall-statistics里的阅读时长要跟3里面的数量对应上 3./group/getClassStatistics、/reading-rank、/textbook-rank-by-student、/teacher-textbook-statistics/student-question-answering-statistics①阅读量都改成小数点后两位②/group/getClassStatistics里的阅读量应该为/reading-rank查询出来的之和

+++ main/src/main/java/com/upc/modular/datastatistics/controller/param/StudentTextbookRankParam.java
@@ -17,5 +17,5 @@ public class StudentTextbookRankParam {
     private String textbookName;
 
     @ApiModelProperty("阅读时长(小时)")
-    private Long readingTime;
+    private double readingTime;
 }
